+++ src/assetbundles/instantanalytics/src/js/InstantAnalytics.js
@@ -1,3 +1,3 @@
 // Import our CSS
-import styles from '../css/app.pcss';
-import icon from '../img/InstantAnalytics-icon.svg';
+import '../css/app.pcss';
+import '../img/InstantAnalytics-icon.svg';